Move update_url_highlight to event::Processor
